add testdata.Path interface
