+++ test/end2end_test.go
@@ -55,6 +55,7 @@ import (
 	"google.golang.org/grpc/status"
 	"google.golang.org/grpc/tap"
 	testpb "google.golang.org/grpc/test/grpc_testing"
+	"google.golang.org/grpc/testdata"
 )
 
 var (
@@ -351,8 +352,6 @@ func (s *testServer) HalfDuplexCall(stream testpb.TestService_HalfDuplexCallServ
 	return nil
 }
 
-const tlsDir = "testdata/"
-
 type env struct {
 	name        string
 	network     string // The type of network such as tcp, unix, etc.
@@ -530,7 +529,7 @@ func (te *test) startServer(ts testpb.TestServiceServer) {
 	}
 	switch te.e.security {
 	case "tls":
-		creds, err := credentials.NewServerTLSFromFile(tlsDir+"server1.pem", tlsDir+"server1.key")
+		creds, err := credentials.NewServerTLSFromFile(testdata.Path("server1.pem"), testdata.Path("server1.key"))
 		if err != nil {
 			te.t.Fatalf("Failed to generate credentials %v", err)
 		}
@@ -605,7 +604,7 @@ func (te *test) clientConn() *grpc.ClientConn {
 	}
 	switch te.e.security {
 	case "tls":
-		creds, err := credentials.NewClientTLSFromFile(tlsDir+"ca.pem", "x.test.youtube.com")
+		creds, err := credentials.NewClientTLSFromFile(testdata.Path("ca.pem"), "x.test.youtube.com")
 		if err != nil {
 			te.t.Fatalf("Failed to load credentials: %v", err)
 		}
